Main develop (#7)
* Plans

* create plans

* create plans

* subscriptions

* subscription migrations

* subscriptions

* subscriptions relations

* create subscription

* subscription created

* adjusts

* adjusts

* more project adjusts

* update subscription

* orders

* orders maps

* orders

* creating orders

* Google maps api

* pricing order

* remove key

* cache server

* update an order

* starting payment pipeline

* Generic Repository

* Generic repository

* Message

* Specific message system

* Hilo

* unity of work

* apply infra pattern

* External services

* adjusts on generi repository

* Refactoring Repository

* Refactoring

* Generic repository

* generic repository

* Messaging (#3)

* testing message system

* disable masstransit

* rabbitmq

* rabbitmq raiz

* messaging full finalized

* unit of work

* Geolocation (#4)

* testing message system

* disable masstransit

* rabbitmq

* rabbitmq raiz

* messaging full finalized

* unit of work

* starting geo

* datetime adjusts

* geolocation done

* Rejections (#5)

* testing message system

* disable masstransit

* rabbitmq

* rabbitmq raiz

* messaging full finalized

* unit of work

* starting geo

* datetime adjusts

* geolocation done

* geo location on redis

* Geolocation feature on redis

* search by geolocation on redis

* Order created domain events

* Log input

* Return Geo space item

* cancel expired order

* Expire old orders

* geolocation ajusts

* reject a ride

* Testing rejections

* Expiration and Rejection

* merge

* ready to implement

+++ Domain/Features/Orders/Enums/AddressType.cs
@@ -1,11 +1,6 @@
-using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Threading.Tasks;
-
 namespace Domain.Enums
 {
-    public enum AddressType : int
+    public enum AddressType : ushort
     {
         StartPoint = 1,
         StoppingPoint = 2,
